Remove custom status exception.
I wanted to the remove the need to have a custom getter on the Pokemon
class to handle bad status types, but I felt that logic should really
live inside of the status class so I had tried overriding the getter in
the EnumMeta, but I'd also want to do the same thing for any other
Enums (eg Pokeballs). This then adds a level of complexity when using
these enums that I'm not sure is needed in that they all raise different
exceptions for invalid key accesses. I may move back to this approach
in the fututure but as of now I'm taking these out.