Change id on Enums to display.
The id was mostly a hack to keep multiple members of the enum of the
same values around, which brought into question why I needed more than
one enum of the same value and it was strictly to keep the display value
so making the display part of the enum value made sense. It also means I
won't have to rely on member.name for display purposes.